moved a lot of stuff to utils, cleaned things up
